Added download_hf.py to download subsets of large datasets from huggingface and modified data.py
download_hf.py is a script that uses parallel workers to download a subset of a huggingface dataset. There were no easy solutions to download a subset of a sharded dataset on huggingface. The implementation is flexible to URL based large datasets or datasets with images directly. It also supports can download metadata.

ImageMetaDataset is a new class in data.py which uses any dataset downloaded with the script or with a similar structure. It will be especially useful for class-conditional generation.